requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.0.0
httpx>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.3
